IS-3979: Do not allow import from /test to /src
