+++ test/components/personkort/PersonkortVisningTest.tsx
@@ -1,5 +1,5 @@
 import React from "react";
-import { expect, describe, it, beforeEach } from "vitest";
+import { beforeEach, describe, expect, it } from "vitest";
 import { PERSONKORTVISNING_TYPE } from "@/konstanter";
 import { QueryClientProvider } from "@tanstack/react-query";
 import { behandlendeEnhetQueryKeys } from "@/data/behandlendeenhet/behandlendeEnhetQueryHooks";
@@ -10,8 +10,10 @@ import { queryClientWithAktivBruker } from "../../testQueryClient";
 import { ARBEIDSTAKER_DEFAULT } from "@/mocks/common/mockConstants";
 import { brukerQueryKeys } from "@/data/navbruker/navbrukerQueryHooks";
 import { brukerinfoMock } from "@/mocks/syfoperson/persondataMock";
-import { daysFromToday } from "../../testUtils";
-import { tilLesbarPeriodeMedArUtenManednavn } from "@/utils/datoUtils";
+import {
+  daysFromToday,
+  tilLesbarPeriodeMedArUtenManednavn,
+} from "@/utils/datoUtils";
 import { behandlendeEnhetMockResponse } from "@/mocks/syfobehandlendeenhet/behandlendeEnhetMock";
 import { PersonkortVisning } from "@/components/personkort/PersonkortVisning";
